로그인/로그아웃 호출 관련 수정
