+++ front_end/src/hooks/api/useLoginWithGoogle.js
@@ -41,6 +41,7 @@ const useLoginWithGoogle = () => {
             // 이메일이 중복되면 아이디 세션스토리지에 저장후 로그인 완료
             sessionStorage.setItem("id", id);
             navigate("/");
+            window.location.reload(); // 페이지 리로드
           }
         })
         .catch((err) => console.log(err));
